# api/api.go
package api

import (
	"net/http"
	"scim-go/model"
	"scim-go/store"

	_ "scim-go/docs"

	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// ScimConfig SCIM配置结构体（从主配置注入）
type ScimConfig struct {
	DefaultSchema string
	GroupSchema   string
	ErrorSchema   string
	ListSchema    string
	APIPath       string
	MaxCount      int
	DefaultCount  int
}

// RouteRegistrar 路由注册器
// 用于模块化注册不同类型的路由
type RouteRegistrar struct {
	r              *gin.Engine
	store          store.Store
	cfg            *ScimConfig
	authToken      string
	swaggerEnabled bool
	swaggerPath    string
}

// NewRouteRegistrar 创建新的路由注册器
func NewRouteRegistrar(r *gin.Engine, s store.Store, cfg *ScimConfig, authToken string, swaggerEnabled bool, swaggerPath string) *RouteRegistrar {
	return &RouteRegistrar{
		r:              r,
		store:          s,
		cfg:            cfg,
		authToken:      authToken,
		swaggerEnabled: swaggerEnabled,
		swaggerPath:    swaggerPath,
	}
}

// RegisterRoutes 注册所有SCIM接口
// 主入口函数，协调所有路由的注册
func RegisterRoutes(r *gin.Engine, s store.Store, cfg *ScimConfig, authToken string, swaggerEnabled bool, swaggerPath string) {
	registrar := NewRouteRegistrar(r, s, cfg, authToken, swaggerEnabled, swaggerPath)

	// 注册全局中间件
	registrar.registerGlobalMiddlewares()

	// 注册Swagger文档路由
	registrar.registerSwaggerRoutes()

	// 注册健康检查路由
	registrar.registerHealthRoutes()

	// 注册服务发现路由
	registrar.registerDiscoveryRoutes()

	// 注册SCIM核心路由
	registrar.registerScimRoutes()
}

// registerGlobalMiddlewares 注册全局中间件
func (reg *RouteRegistrar) registerGlobalMiddlewares() {
	// 配置CORS中间件
	corsConfig := cors.DefaultConfig()
	corsConfig.AllowOrigins = []string{"*"}
	corsConfig.AllowMethods = []string{"GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"}
	corsConfig.AllowHeaders = []string{"Origin", "Content-Type", "Accept", "Authorization"}
	corsConfig.AllowCredentials = true
	reg.r.Use(cors.New(corsConfig))

	// 注册性能监控中间件
	reg.r.Use(PerformanceMonitor())

	// 注册恢复中间件
	reg.r.Use(Recovery())
}

// registerSwaggerRoutes 注册Swagger文档路由
func (reg *RouteRegistrar) registerSwaggerRoutes() {
	if !reg.swaggerEnabled {
		return
	}

	// 处理所有 Swagger 相关路径
	reg.r.GET(reg.swaggerPath+"/*any", func(c *gin.Context) {
		// 获取路径参数
		path := c.Param("any")
		// 如果路径为空或只是斜杠，重定向到 index.html
		if path == "" || path == "/" {
			c.Redirect(http.StatusMovedPermanently, reg.swaggerPath+"/index.html")
			return
		}
		// 否则使用 Swagger 处理器
		ginSwagger.WrapHandler(swaggerFiles.Handler)(c)
	})

	// 处理根路径请求（如 /swagger）
	reg.r.GET(reg.swaggerPath, func(c *gin.Context) {
		c.Redirect(http.StatusMovedPermanently, reg.swaggerPath+"/index.html")
	})
}

// registerHealthRoutes 注册健康检查路由
func (reg *RouteRegistrar) registerHealthRoutes() {
	// 健康检查（无需认证）
	reg.r.GET("/health", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{
			"status":  "ok",
			"service": "scim-server",
			"version": "1.0.0",
		})
	})

	// 就绪检查（无需认证）
	reg.r.GET("/ready", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{
			"status": "ready",
		})
	})
}

// registerDiscoveryRoutes 注册服务发现路由
func (reg *RouteRegistrar) registerDiscoveryRoutes() {
	// SCIM根路径（服务发现，无需认证）
	reg.r.GET(reg.cfg.APIPath, func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{
			"schemas": []string{"urn:ietf:params:scim:api:messages:2.0:ServiceProviderConfig"},
			"patch":   gin.H{"supported": true},
			"filter":  gin.H{"supported": true, "maxResults": reg.cfg.MaxCount},
			"sort":    gin.H{"supported": true},
			"pagination": gin.H{
				"supported":       true,
				"maxItemsPerPage": reg.cfg.MaxCount,
			},
			"resources": gin.H{
				"User":  gin.H{"endpoint": reg.cfg.APIPath + "/Users"},
				"Group": gin.H{"endpoint": reg.cfg.APIPath + "/Groups"},
			},
		})
	})
}

// registerScimRoutes 注册SCIM核心路由
func (reg *RouteRegistrar) registerScimRoutes() {
	// ServiceProviderConfig 端点（无需认证，符合 RFC 7644）
	reg.r.GET(reg.cfg.APIPath+"/ServiceProviderConfig", func(c *gin.Context) {
		config := model.GetServiceProviderConfig(reg.cfg.MaxCount)
		c.JSON(http.StatusOK, config)
	})

	// SCIM核心组（需要认证+参数绑定+分页）
	scimGroup := reg.r.Group(reg.cfg.APIPath)
	scimGroup.Use(Auth(reg.authToken))
	scimGroup.Use(BindQuery())
	scimGroup.Use(Pagination(reg.cfg.DefaultCount, reg.cfg.MaxCount))

	// 注册ResourceTypes路由
	reg.registerResourceTypesRoutes(scimGroup)

	// 注册Schemas路由
	reg.registerSchemasRoutes(scimGroup)

	// 注册User和Group路由
	reg.registerUserRoutes(scimGroup)
	reg.registerGroupRoutes(scimGroup)

	// 注册自定义资源类型路由
	reg.registerCustomResourceTypesRoutes(scimGroup)

	// 注册自定义资源路由（放在最后，避免影响其他路由）
	reg.registerCustomResourcesRoutes(scimGroup)
}

// registerResourceTypesRoutes 注册ResourceTypes路由
func (reg *RouteRegistrar) registerResourceTypesRoutes(scimGroup *gin.RouterGroup) {
	// ResourceTypes 端点（需要认证）
	scimGroup.GET("/ResourceTypes", func(c *gin.Context) {
		// 获取完整的基础URL
		baseURL := getBaseURL(c)

		// 构建完整的ResourceType对象
		userResourceType := model.GetUserResourceType()
		userResourceType.Endpoint = baseURL + reg.cfg.APIPath + "/Users"
		userResourceType.Meta = &model.Meta{
			ResourceType: "ResourceType",
			Location:     baseURL + reg.cfg.APIPath + "/ResourceTypes/User",
		}

		groupResourceType := model.GetGroupResourceType()
		groupResourceType.Endpoint = baseURL + reg.cfg.APIPath + "/Groups"
		groupResourceType.Meta = &model.Meta{
			ResourceType: "ResourceType",
			Location:     baseURL + reg.cfg.APIPath + "/ResourceTypes/Group",
		}

		// 获取自定义资源类型
		customResourceTypes, err := reg.store.ListCustomResourceTypes()
		if err != nil {
			ErrorHandler(c, err, http.StatusInternalServerError, "internalError")
			return
		}

		// 构建完整的自定义ResourceType对象
		for i := range customResourceTypes {
			crt := &customResourceTypes[i]
			crt.Endpoint = baseURL + reg.cfg.APIPath + crt.Endpoint
			crt.Meta = &model.Meta{
				ResourceType: "ResourceType",
				Location:     baseURL + reg.cfg.APIPath + "/ResourceTypes/" + crt.ID,
			}
		}

		// 合并所有资源类型
		resourceTypes := []model.ResourceType{
			*userResourceType,
			*groupResourceType,
		}
		// 将自定义资源类型转换为ResourceType并添加到列表
		for _, crt := range customResourceTypes {
			resourceType := model.ResourceType{
				Schemas:          crt.Schemas,
				ID:               crt.ID,
				Name:             crt.Name,
				Endpoint:         crt.Endpoint,
				Description:      crt.Description,
				Schema:           crt.Schema,
				SchemaExtensions: crt.SchemaExtensions,
				Meta:             crt.Meta,
			}
			resourceTypes = append(resourceTypes, resourceType)
		}

		c.JSON(http.StatusOK, model.ListResponse{
			Schemas:      []string{model.ListSchema.String()},
			TotalResults: len(resourceTypes),
			Resources:    resourceTypes,
		})
	})

	scimGroup.GET("/ResourceTypes/:id", func(c *gin.Context) {
		id := c.Param("id")
		// 获取完整的基础URL
		baseURL := getBaseURL(c)

		switch id {
		case "User":
			resourceType := model.GetUserResourceType()
			resourceType.Endpoint = reg.cfg.APIPath + "/Users"
			resourceType.Meta = &model.Meta{
				ResourceType: "ResourceType",
				Location:     baseURL + reg.cfg.APIPath + "/ResourceTypes/User",
			}
			c.JSON(http.StatusOK, resourceType)
		case "Group":
			resourceType := model.GetGroupResourceType()
			resourceType.Endpoint = reg.cfg.APIPath + "/Groups"
			resourceType.Meta = &model.Meta{
				ResourceType: "ResourceType",
				Location:     baseURL + reg.cfg.APIPath + "/ResourceTypes/Group",
			}
			c.JSON(http.StatusOK, resourceType)
		default:
			// 尝试获取自定义资源类型
			crt, err := reg.store.GetCustomResourceType(id)
			if err != nil {
				ErrorHandler(c, err, http.StatusNotFound, "notFound")
				return
			}
			// 构建完整的自定义ResourceType对象
			crt.Endpoint = baseURL + reg.cfg.APIPath + crt.Endpoint
			crt.Meta = &model.Meta{
				ResourceType: "ResourceType",
				Location:     baseURL + reg.cfg.APIPath + "/ResourceTypes/" + crt.ID,
			}
			// 转换为ResourceType并返回
			resourceType := model.ResourceType{
				Schemas:          crt.Schemas,
				ID:               crt.ID,
				Name:             crt.Name,
				Endpoint:         crt.Endpoint,
				Description:      crt.Description,
				Schema:           crt.Schema,
				SchemaExtensions: crt.SchemaExtensions,
				Meta:             crt.Meta,
			}
			c.JSON(http.StatusOK, resourceType)
		}
	})
}

// registerCustomResourceTypesRoutes 注册自定义资源类型的CRUD路由
func (reg *RouteRegistrar) registerCustomResourceTypesRoutes(scimGroup *gin.RouterGroup) {
	// 自定义资源类型管理端点
	customResourceTypesGroup := scimGroup.Group("/CustomResourceTypes")

	// 注册自定义资源类型的CRUD路由
	customResourceTypesGroup.POST("", reg.createCustomResourceType)
	customResourceTypesGroup.GET("/:id", reg.getCustomResourceType)
	customResourceTypesGroup.GET("", reg.listCustomResourceTypes)
	customResourceTypesGroup.PUT("/:id", reg.updateCustomResourceType)
	customResourceTypesGroup.DELETE("/:id", reg.deleteCustomResourceType)
}

// registerCustomResourcesRoutes 注册自定义资源的CRUD路由
func (reg *RouteRegistrar) registerCustomResourcesRoutes(scimGroup *gin.RouterGroup) {
	// 动态注册自定义资源路由 - 使用参数路由而不是通配符路由

	// 注册自定义资源的CRUD路由
	scimGroup.GET("/:resourceType", reg.listCustomResources)
	scimGroup.POST("/:resourceType", reg.createCustomResource)
	scimGroup.GET("/:resourceType/:resourceID", reg.getCustomResource)
	scimGroup.PUT("/:resourceType/:resourceID", reg.updateCustomResource)
	scimGroup.PATCH("/:resourceType/:resourceID", reg.patchCustomResource)
	scimGroup.DELETE("/:resourceType/:resourceID", reg.deleteCustomResource)
}

// getBaseURL 获取完整的基础URL
func getBaseURL(c *gin.Context) string {
	protocol := "http"
	if c.Request.TLS != nil {
		protocol = "https"
	}
	return protocol + "://" + c.Request.Host
}

// registerSchemasRoutes 注册Schemas路由
func (reg *RouteRegistrar) registerSchemasRoutes(scimGroup *gin.RouterGroup) {
	// Schemas 端点（需要认证）
	scimGroup.GET("/Schemas", func(c *gin.Context) {
		schemas := []model.Schema{
			*model.GetUserSchema(),
			*model.GetGroupSchema(),
			*model.GetEnterpriseUserExtensionSchema(),
		}
		c.JSON(http.StatusOK, model.ListResponse{
			Schemas:      []string{model.ListSchema.String()},
			TotalResults: len(schemas),
			Resources:    schemas,
		})
	})

	scimGroup.GET("/Schemas/:id", func(c *gin.Context) {
		id := c.Param("id")
		switch id {
		case model.UserSchema.String():
			c.JSON(http.StatusOK, model.GetUserSchema())
		case model.GroupSchema.String():
			c.JSON(http.StatusOK, model.GetGroupSchema())
		case model.EnterpriseUserSchema.String():
			c.JSON(http.StatusOK, model.GetEnterpriseUserExtensionSchema())
		default:
			ErrorHandler(c, model.ErrNotFound, http.StatusNotFound, "notFound")
		}
	})
}

// registerUserRoutes 注册User路由
func (reg *RouteRegistrar) registerUserRoutes(scimGroup *gin.RouterGroup) {
	// 初始化用户处理器
	userHandler := NewUserHandlers(reg.store, &model.ScimConfig{
		DefaultSchema: reg.cfg.DefaultSchema,
		GroupSchema:   reg.cfg.GroupSchema,
		ErrorSchema:   reg.cfg.ErrorSchema,
		ListSchema:    reg.cfg.ListSchema,
		APIPath:       reg.cfg.APIPath,
	})

	// 注册User接口（全方法：GET/POST/PUT/PATCH/DELETE）
	userGroup := scimGroup.Group("/Users")
	userGroup.GET("", userHandler.ListUsers)
	userGroup.GET("/:id", userHandler.GetUser)
	userGroup.POST("", userHandler.CreateUser)
	userGroup.PUT("/:id", userHandler.UpdateUser)
	userGroup.PATCH("/:id", userHandler.PatchUser)
	userGroup.DELETE("/:id", userHandler.DeleteUser)
}

// registerGroupRoutes 注册Group路由
func (reg *RouteRegistrar) registerGroupRoutes(scimGroup *gin.RouterGroup) {
	// 初始化组处理器
	groupHandler := NewGroupHandlers(reg.store, &model.ScimConfig{
		DefaultSchema: reg.cfg.DefaultSchema,
		GroupSchema:   reg.cfg.GroupSchema,
		ErrorSchema:   reg.cfg.ErrorSchema,
		ListSchema:    reg.cfg.ListSchema,
		APIPath:       reg.cfg.APIPath,
	})

	// 注册Group接口（全方法：GET/POST/PUT/PATCH/DELETE）
	groupGroup := scimGroup.Group("/Groups")
	groupGroup.GET("", groupHandler.ListGroups)
	groupGroup.GET("/:id", groupHandler.GetGroup)
	groupGroup.POST("", groupHandler.CreateGroup)
	groupGroup.PUT("/:id", groupHandler.UpdateGroup)
	groupGroup.PATCH("/:id", groupHandler.PatchGroup)
	groupGroup.DELETE("/:id", groupHandler.DeleteGroup)

	// 注册Group成员管理接口
	groupGroup.GET("/:id/members", groupHandler.GetGroupMembers)
	groupGroup.POST("/:id/members", groupHandler.AddMembersToGroup)
	groupGroup.DELETE("/:id/members/:userId", groupHandler.RemoveMemberFromGroup)
}