# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SCIM (System for Cross-domain Identity Management) 2.0 实现，提供用户和组管理功能",
        "title": "SCIM 2.0 API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "shoy160",
            "email": "shoy160@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/scim/v2",
    "paths": {
        "/Groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取组列表，支持分页、过滤和属性选择",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "列出所有组",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "起始索引",
                        "name": "startIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "过滤条件",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "要返回的属性列表，逗号分隔",
                        "name": "attributes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "要排除的属性列表，逗号分隔",
                        "name": "excludedAttributes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "组列表",
                        "schema": {
                            "$ref": "#/definitions/model.ListResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新组，需要提供组显示名称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "创建组",
                "parameters": [
                    {
                        "description": "组信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "组已存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/Groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据组ID获取组详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "获取单个组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要返回的属性列表，逗号分隔",
                        "name": "attributes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "要排除的属性列表，逗号分隔",
                        "name": "excludedAttributes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "组信息",
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "组不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "全量替换组信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "全量更新组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "组信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "组不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据组ID删除组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "删除组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功"
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "组不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "使用补丁操作更新组属性",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "补丁更新组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "补丁操作",
                        "name": "patch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "组不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/Groups/{id}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定组的成员列表，支持分页和成员类型过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "获取组成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "起始索引（从1开始）",
                        "name": "startIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "成员类型过滤（User或Group）",
                        "name": "memberType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "组成员列表",
                        "schema": {
                            "$ref": "#/definitions/model.ListResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "组不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将一个或多个成员添加到指定组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "添加成员到组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "单个成员信息",
                        "name": "member",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Member"
                        }
                    },
                    {
                        "description": "成员信息数组",
                        "name": "members",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Member"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "添加成功",
                        "schema": {
                            "$ref": "#/definitions/model.Group"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "组或成员不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/Groups/{id}/members/{userId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将成员从指定组中移除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "从组中移除成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "成员ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "移除成功"
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "组或成员不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail 错误详情",
                    "type": "string"
                },
                "schemas": {
                    "description": "Schemas SCIM错误Schema",
                    "type": "string"
                },
                "scimType": {
                    "description": "ScimType SCIM错误类型",
                    "type": "string"
                },
                "status": {
                    "description": "Status HTTP状态码",
                    "type": "integer"
                }
            }
        },
...
        "model.ListResponse": {
            "type": "object",
            "properties": {
                "Resources": {
                    "description": "Resources 资源列表"
                },
                "itemsPerPage": {
                    "description": "ItemsPerPage 每页项目数",
                    "type": "integer"
                },
                "schemas": {
                    "description": "Schemas SCIM Schema标识符数组",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "startIndex": {
                    "description": "StartIndex 起始索引（从1开始）",
                    "type": "integer"
                },
                "totalResults": {
                    "description": "TotalResults 总结果数量",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请输入 \"Bearer {token}\" 格式的认证令牌",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}